refactorng(components):
separated entites one per file
separated utility functions by concern
